Exception handling - icsmv001

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 @Service
@@ -22,11 +26,12 @@ public class ProductService {
 	//metodo de busca no banco de dados
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional <Product> result = repository.findById(id);
-		Product product = result.get();
-		ProductDTO dto = new ProductDTO(product);
-		return dto;
+ 
 		
+		//tratamento de retorno de erro, colocando o retorno tradado.
+		Product product = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso nao encontrado"));
+		       return new ProductDTO(product);
 	}
 	
 	
@@ -51,35 +56,63 @@ public class ProductService {
 		}		
 	
 		
+		
+		
+		
 		//metodo alterar item  (PUT)
 				@Transactional
 				public ProductDTO update(Long id,ProductDTO dto) {
 					
-					Product entity = repository.getReferenceById(id);
-
-                    copyDtoToEntity(dto,entity);
+					try {
 					
-				    entity = repository.save(entity);
-					return new ProductDTO(entity);
+						Product entity = repository.getReferenceById(id);
+	                    copyDtoToEntity(dto,entity);
+					    entity = repository.save(entity);
+						return new ProductDTO(entity);
+						
+					}
+					
+					catch(EntityNotFoundException e) { 
+						throw new  ResourceNotFoundException("Recurso nao encontrado");
+						
+					}
+					
+					
 				}
 
 				
+				private void copyDtoToEntity(ProductDTO dto, Product entity) {
+					entity.setName(dto.getName());
+					entity.setDescription(dto.getDescription());
+					entity.setPrice(dto.getPrice());
+					entity.setImgUrl(dto.getDescription());
+					
+				}	
+				
+				
+				
 				
 				//METODO DELETE
-				@Transactional
+				@Transactional(propagation = Propagation.SUPPORTS)
 				public void delete(Long id) {
-				repository.deleteById(id);
+					
+					if (!repository.existsById(id)) {
+						throw new ResourceNotFoundException("Recurso nao encontrado");
+					}
+					
+					try {
+						repository.deleteById(id);
+					}
+					catch (DataIntegrityViolationException e) {
+						throw new DatabaseException("Falha de integracao referencial");
+					}
+					
+					
+				
 				}		
 				
 		
 
-		private void copyDtoToEntity(ProductDTO dto, Product entity) {
-			entity.setName(dto.getName());
-			entity.setDescription(dto.getDescription());
-			entity.setPrice(dto.getPrice());
-			entity.setImgUrl(dto.getDescription());
-			
-		}
